+++ src/oci/core/models/instance_configuration_launch_instance_shape_config_details.py
@@ -35,16 +35,23 @@ class InstanceConfigurationLaunchInstanceShapeConfigDetails(object):
             The value to assign to the ocpus property of this InstanceConfigurationLaunchInstanceShapeConfigDetails.
         :type ocpus: float
 
+        :param memory_in_gbs:
+            The value to assign to the memory_in_gbs property of this InstanceConfigurationLaunchInstanceShapeConfigDetails.
+        :type memory_in_gbs: float
+
         """
         self.swagger_types = {
-            'ocpus': 'float'
+            'ocpus': 'float',
+            'memory_in_gbs': 'float'
         }
 
         self.attribute_map = {
-            'ocpus': 'ocpus'
+            'ocpus': 'ocpus',
+            'memory_in_gbs': 'memoryInGBs'
         }
 
         self._ocpus = None
+        self._memory_in_gbs = None
 
     @property
     def ocpus(self):
@@ -69,6 +76,30 @@ class InstanceConfigurationLaunchInstanceShapeConfigDetails(object):
         :type: float
         """
         self._ocpus = ocpus
+
+    @property
+    def memory_in_gbs(self):
+        """
+        Gets the memory_in_gbs of this InstanceConfigurationLaunchInstanceShapeConfigDetails.
+        The total amount of memory available to the instance, in gigabytes.
+
+
+        :return: The memory_in_gbs of this InstanceConfigurationLaunchInstanceShapeConfigDetails.
+        :rtype: float
+        """
+        return self._memory_in_gbs
+
+    @memory_in_gbs.setter
+    def memory_in_gbs(self, memory_in_gbs):
+        """
+        Sets the memory_in_gbs of this InstanceConfigurationLaunchInstanceShapeConfigDetails.
+        The total amount of memory available to the instance, in gigabytes.
+
+
+        :param memory_in_gbs: The memory_in_gbs of this InstanceConfigurationLaunchInstanceShapeConfigDetails.
+        :type: float
+        """
+        self._memory_in_gbs = memory_in_gbs
 
     def __repr__(self):
         return formatted_flat_dict(self)