+++ src/oci/database/models/autonomous_database_summary.py
@@ -163,6 +163,14 @@ class AutonomousDatabaseSummary(object):
     #: This constant has a value of "MANUAL"
     REFRESHABLE_MODE_MANUAL = "MANUAL"
 
+    #: A constant which can be used with the permission_level property of a AutonomousDatabaseSummary.
+    #: This constant has a value of "RESTRICTED"
+    PERMISSION_LEVEL_RESTRICTED = "RESTRICTED"
+
+    #: A constant which can be used with the permission_level property of a AutonomousDatabaseSummary.
+    #: This constant has a value of "UNRESTRICTED"
+    PERMISSION_LEVEL_UNRESTRICTED = "UNRESTRICTED"
+
     def __init__(self, **kwargs):
         """
         Initializes a new AutonomousDatabaseSummary object with values from keyword arguments.
@@ -360,6 +368,12 @@ class AutonomousDatabaseSummary(object):
             The value to assign to the source_id property of this AutonomousDatabaseSummary.
         :type source_id: str
 
+        :param permission_level:
+            The value to assign to the permission_level property of this AutonomousDatabaseSummary.
+            Allowed values for this property are: "RESTRICTED", "UNRESTRICTED", 'UNKNOWN_ENUM_VALUE'.
+            Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
+        :type permission_level: str
+
         :param time_of_last_switchover:
             The value to assign to the time_of_last_switchover property of this AutonomousDatabaseSummary.
         :type time_of_last_switchover: datetime
@@ -430,6 +444,7 @@ class AutonomousDatabaseSummary(object):
             'refreshable_status': 'str',
             'refreshable_mode': 'str',
             'source_id': 'str',
+            'permission_level': 'str',
             'time_of_last_switchover': 'datetime',
             'time_of_last_failover': 'datetime',
             'is_data_guard_enabled': 'bool',
@@ -483,6 +498,7 @@ class AutonomousDatabaseSummary(object):
             'refreshable_status': 'refreshableStatus',
             'refreshable_mode': 'refreshableMode',
             'source_id': 'sourceId',
+            'permission_level': 'permissionLevel',
             'time_of_last_switchover': 'timeOfLastSwitchover',
             'time_of_last_failover': 'timeOfLastFailover',
             'is_data_guard_enabled': 'isDataGuardEnabled',
@@ -535,6 +551,7 @@ class AutonomousDatabaseSummary(object):
         self._refreshable_status = None
         self._refreshable_mode = None
         self._source_id = None
+        self._permission_level = None
         self._time_of_last_switchover = None
         self._time_of_last_failover = None
         self._is_data_guard_enabled = None
@@ -1745,6 +1762,36 @@ class AutonomousDatabaseSummary(object):
         :type: str
         """
         self._source_id = source_id
+
+    @property
+    def permission_level(self):
+        """
+        Gets the permission_level of this AutonomousDatabaseSummary.
+        The Autonomous Database permission level. Restricted mode allows access only to admin users.
+
+        Allowed values for this property are: "RESTRICTED", "UNRESTRICTED", 'UNKNOWN_ENUM_VALUE'.
+        Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
+
+
+        :return: The permission_level of this AutonomousDatabaseSummary.
+        :rtype: str
+        """
+        return self._permission_level
+
+    @permission_level.setter
+    def permission_level(self, permission_level):
+        """
+        Sets the permission_level of this AutonomousDatabaseSummary.
+        The Autonomous Database permission level. Restricted mode allows access only to admin users.
+
+
+        :param permission_level: The permission_level of this AutonomousDatabaseSummary.
+        :type: str
+        """
+        allowed_values = ["RESTRICTED", "UNRESTRICTED"]
+        if not value_allowed_none_or_none_sentinel(permission_level, allowed_values):
+            permission_level = 'UNKNOWN_ENUM_VALUE'
+        self._permission_level = permission_level
 
     @property
     def time_of_last_switchover(self):