+++ src/oci/key_management/models/create_key_details.py
@@ -13,6 +13,14 @@ class CreateKeyDetails(object):
     CreateKeyDetails model.
     """
 
+    #: A constant which can be used with the protection_mode property of a CreateKeyDetails.
+    #: This constant has a value of "HSM"
+    PROTECTION_MODE_HSM = "HSM"
+
+    #: A constant which can be used with the protection_mode property of a CreateKeyDetails.
+    #: This constant has a value of "SOFTWARE"
+    PROTECTION_MODE_SOFTWARE = "SOFTWARE"
+
     def __init__(self, **kwargs):
         """
         Initializes a new CreateKeyDetails object with values from keyword arguments.
@@ -38,13 +46,19 @@ class CreateKeyDetails(object):
             The value to assign to the key_shape property of this CreateKeyDetails.
         :type key_shape: KeyShape
 
+        :param protection_mode:
+            The value to assign to the protection_mode property of this CreateKeyDetails.
+            Allowed values for this property are: "HSM", "SOFTWARE"
+        :type protection_mode: str
+
         """
         self.swagger_types = {
             'compartment_id': 'str',
             'defined_tags': 'dict(str, dict(str, object))',
             'display_name': 'str',
             'freeform_tags': 'dict(str, str)',
-            'key_shape': 'KeyShape'
+            'key_shape': 'KeyShape',
+            'protection_mode': 'str'
         }
 
         self.attribute_map = {
@@ -52,7 +66,8 @@ class CreateKeyDetails(object):
             'defined_tags': 'definedTags',
             'display_name': 'displayName',
             'freeform_tags': 'freeformTags',
-            'key_shape': 'keyShape'
+            'key_shape': 'keyShape',
+            'protection_mode': 'protectionMode'
         }
 
         self._compartment_id = None
@@ -60,6 +75,7 @@ class CreateKeyDetails(object):
         self._display_name = None
         self._freeform_tags = None
         self._key_shape = None
+        self._protection_mode = None
 
     @property
     def compartment_id(self):
@@ -194,6 +210,46 @@ class CreateKeyDetails(object):
         :type: KeyShape
         """
         self._key_shape = key_shape
+
+    @property
+    def protection_mode(self):
+        """
+        Gets the protection_mode of this CreateKeyDetails.
+        The key's protection mode indicates how the key persists and where cryptographic operations that use the key are performed.
+        A protection mode of `HSM` means that the key persists on a hardware security module (HSM) and all cryptographic operations are performed inside
+        the HSM. A protection mode of `SOFTWARE` means that the key persists on the server, protected by the vault's RSA wrapping key which persists
+        on the HSM. All cryptographic operations that use a key with a protection mode of `SOFTWARE` are performed on the server. By default,
+        a key's protection mode is set to `HSM`. You can't change a key's protection mode after the key is created or imported.
+
+        Allowed values for this property are: "HSM", "SOFTWARE"
+
+
+        :return: The protection_mode of this CreateKeyDetails.
+        :rtype: str
+        """
+        return self._protection_mode
+
+    @protection_mode.setter
+    def protection_mode(self, protection_mode):
+        """
+        Sets the protection_mode of this CreateKeyDetails.
+        The key's protection mode indicates how the key persists and where cryptographic operations that use the key are performed.
+        A protection mode of `HSM` means that the key persists on a hardware security module (HSM) and all cryptographic operations are performed inside
+        the HSM. A protection mode of `SOFTWARE` means that the key persists on the server, protected by the vault's RSA wrapping key which persists
+        on the HSM. All cryptographic operations that use a key with a protection mode of `SOFTWARE` are performed on the server. By default,
+        a key's protection mode is set to `HSM`. You can't change a key's protection mode after the key is created or imported.
+
+
+        :param protection_mode: The protection_mode of this CreateKeyDetails.
+        :type: str
+        """
+        allowed_values = ["HSM", "SOFTWARE"]
+        if not value_allowed_none_or_none_sentinel(protection_mode, allowed_values):
+            raise ValueError(
+                "Invalid value for `protection_mode`, must be None or one of {0}"
+                .format(allowed_values)
+            )
+        self._protection_mode = protection_mode
 
     def __repr__(self):
         return formatted_flat_dict(self)